ProjMgtService-Refactoring: Created new services for UserStories and Tasks, refactored the packages for the services.
