+++ src/main/java/com/umssonline/proymgt/controllers/ProjectRestController.java
@@ -3,10 +3,9 @@ package com.umssonline.proymgt.controllers;
 import com.umssonline.proymgt.models.dto.project.CreateProjectDto;
 import com.umssonline.proymgt.models.dto.project.UpdateProjectDto;
 import com.umssonline.proymgt.models.dto.sprint.CreateSprintDto;
-import com.umssonline.proymgt.models.entity.Backlog;
 import com.umssonline.proymgt.models.entity.Project;
 import com.umssonline.proymgt.models.entity.Sprint;
-import com.umssonline.proymgt.services.ProjectServiceImpl;
+import com.umssonline.proymgt.services.impl.ProjectServiceImpl;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
