+++ src/main/java/com/umssonline/proymgt/services/api/CrudService.java
@@ -1,4 +1,4 @@
-package com.umssonline.proymgt.services;
+package com.umssonline.proymgt.services.api;
 
 import java.io.Serializable;
