+++ src/main/java/com/umssonline/proymgt/services/api/SprintService.java
@@ -1,4 +1,4 @@
-package com.umssonline.proymgt.services;
+package com.umssonline.proymgt.services.api;
 
 import com.umssonline.proymgt.models.entity.Sprint;
