+++ src/main/java/com/umssonline/proymgt/services/impl/ProjectServiceImpl.java
@@ -1,4 +1,4 @@
-package com.umssonline.proymgt.services;
+package com.umssonline.proymgt.services.impl;
 
 import com.umssonline.proymgt.models.entity.Backlog;
 import com.umssonline.proymgt.models.entity.Project;
@@ -6,6 +6,7 @@ import com.umssonline.proymgt.models.entity.Sprint;
 import com.umssonline.proymgt.repositories.BacklogRepository;
 import com.umssonline.proymgt.repositories.ProjectRepository;
 import com.umssonline.proymgt.repositories.SprintRepository;
+import com.umssonline.proymgt.services.api.ProjectService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
