+++ src/main/java/com/umssonline/proymgt/services/impl/UserService.java
@@ -1,4 +1,4 @@
-package com.umssonline.proymgt.services;
+package com.umssonline.proymgt.services.impl;
 
 public class UserService {
 }
